refactor: SmsService와 SecureKeypadService를 완전히 AuthenticationApiService로 통합
