+++ src/main/java/com/swifty/bank/server/core/domain/keypad/service/SecureKeypadService.java
@@ -1,0 +1,7 @@
+package com.swifty.bank.server.core.domain.keypad.service;
+
+import com.swifty.bank.server.core.domain.keypad.service.dto.SecureKeypadDto;
+
+public interface SecureKeypadService {
+    SecureKeypadDto createSecureKeypad();
+}
